% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modgr.R
\name{modgr}
\alias{modgr}
\title{Moda de datos agrupados}
\usage{
modgr(x, breaks = 100)
}
\arguments{
\item{x}{un elemento o vector numérico}

\item{breaks}{número de intervalos que se deben crear para dividir el vector \code{x}.}
}
\description{
Esta función calcula la moda de datos agrupados del vector \code{x}.
}
\details{
La función recibe un vector \code{x} y un número \code{breaks},
divide el vector en el número de intervalos y calcula la moda.
}
\examples{
modgr(rnorm(1000, 20, 15))
modgr(rnorm(1000, 20, 15), 100)
}
